start migration to manifest v3
