+++ manifest.json
@@ -7,7 +7,7 @@
 	},
 	"name": "ETHSBell",
 	"version": "2.4.1",
-	"manifest_version": 2,
+	"manifest_version": 3,
 	"description": "A Browser Extension version of ETHSBell",
 	"homepage_url": "https://ethsbell.app",
 	"icons": {
@@ -15,17 +15,15 @@
 		"48": "icons/icon48.png",
 		"128": "icons/icon128.png"
 	},
-	"background": {
-		"scripts": [
-			"src/helpers.js",
-			"background.js"
-		]
-	},
+  	"background": {
+    	"service_worker": "background.js",
+    	"type": "module"
+  	},
 	"options_ui": {
 		"open_in_tab": false,
 		"page": "/options/index.html"
 	},
-	"browser_action": {
+	"action": {
 		"default_icon": "icons/icon128.png",
 		"default_title": "ETHSBell",
 		"default_popup": "src/browser_action/browser_action.html"
